Removed top level invocations of feature.*
Summary:
When I make buck2 builds use my logic by default in D38135758, I get the following error:
```
$ buck2 build //metalos/initrd:
BUILD FAILED
error: Unable to access module internals. This could be due to accessing it in the context of interpreting a .bzl file.
  --> fbcode/antlir/bzl2/feature_rule.bzl:
   |
   |     if not native.rule_exists(target_name):
```
This diff wraps the lists `​​dbus`, `debug`, and `users` in functions so that this error doesn't occur.

Reviewed By: naveedgol

Differential Revision: D38051566

fbshipit-source-id: 34134f369dd2a6a62277037eab94402d97293213