Fixed bugs in feature.install_buck_runnable
Summary:
I caught two bugs:
1. In the buck2 `feature.install_buck_runnable`, I was missing an argument, so `runs_in_build_steps_causes_slow_rebuilds` was always getting set to `False`.
2. In the buck1 and buck2 `feature.install_buck_runnable`, if you had a target that outputs a directory, and you wanted to install more than one buck runnable artifact contained in that directory, only the first artifact would end up getting installed, and it would be installed repeatedly in place of the artifacts. For example, if you had something like:
```
feature.install_buck_runnable(
    image.source(
        ":dir-print-ok",
        path = "subdir/print-ok",
    ),
    "/foo",
),

feature.install_buck_runnable(
    image.source(
        ":dir-print-ok",
        path = "subdir/print-ok-again",
    ),
    "/bar",
)
```
The artifact at `subdir/print-ok` would be wrapped and installed properly. However, the artifact at `subdir/print-ok-again` would not be wrapped because its wrapped target name would be the same as that of `subdir/print-ok`. This is because `wrap_target` only considers the target name when wrapping a target, and not the path within the target, making the target names overlap. The result of this was that the artifact at `subdir/print-ok` was installed at `/bar` instead of the artifact at `subdir/print-ok-again`. I fixed this issue by adding the path to the wrap suffix, removing the overlap of the target names. This probably was never discovered because it would take a very specific set of circumstances for this to be a problem, but I noticed it while making targets use the buck2 logic. It becomes more likely to be a problem with the buck2 logic because we are wrapping the target every time we call `feature.install`. The actual change was one line, and the rest of the lines are just the tests being updated.

Reviewed By: naveedgol

Differential Revision: D38074563

fbshipit-source-id: 74b459225ab1f71fd59ea735b9840dce19c35168