Made rpm public key destination folder equal in buck1 and buck2
Summary: Previously, `dest` was different in Buck1 and Buck2, resulting in different target names, so I changed how the name is determined. Now, if `src` is a string, it just gets hashed like before, but if it's a shape (which it always will be if it's not a string), it gets hashed with `shape.hash`.

Reviewed By: naveedgol

Differential Revision: D38052708

fbshipit-source-id: b9265d774a5ec48b8a9160f879a101e2a0bc6e71