Changed buck2 rpms.bzl to pass struct to feature.new
Summary:
Previously, calls to rpm actions would create a new target that depended on all of the necessary version sets. However, this forced changes to the `rpms_install` API because in the old logic, version sets targets were filtered in `feature.new` before being added as dependencies.

I fixed this problem by making calls to `feature.rpms_x` not declare a rule but instead return a struct containing rpm items and version set dependencies that can be filtered in calls to `feature.new`. This makes it possible to make existing calls to `rpms_install` use provider logic without requiring changes to the API.

Differential Revision: D38207451

fbshipit-source-id: 3b93d023e2e7b50787e1a4d8ae06cf7b6b9f014d