Fixed buck2 feature.tarball test case
Summary: vmagro had removed generator from image_source and the `feature.tarball` test image, but it hadn't been removed from the buck2 python test for `feature.tarball`, causing the test to break. This diff fixes that. I think it wasn't showing up in CI because the test is "unknown_quality" for some reason.

Reviewed By: naveedgol

Differential Revision: D38437284

fbshipit-source-id: 6467f37b1987a39c3aa1e5552ea76bd40b3e16b2