Made buck1 compile_image_features and feature.new mirror buck2 dep graph
Summary: In buck2 logic, child layer features depend on their parent layer's feature, so I added that dependency to the buck1 versions of `compile_image_features` and `feature.new`. I also had to remove the query from the hash in the `deps-targets-to-outputs` target because the queries in the buck1 and buck2 logic are different, meaning that the resulting target name is also different. I don't think having the query in the hash is necessary anyway.

Differential Revision: D38020587

fbshipit-source-id: 37dd958d9147fd92ddc6543f5247866e582787ea