Added calls to dummy_rule to generic buck1 feature macros
Summary:
Added dummy rule calls to most of the buck1 feature macros to mirror buck2 dep graph.

`is_buck2` flag will be replaced by calls to `is_buck2()` at the top of the stack, but is necessary here because at this point in the stack, buck2 builds still use the buck1 logic.

Differential Revision: D38362516

fbshipit-source-id: eec69e1b705d948ba6da13d71bcd4aa82abddbfa