Add a MetaKeyValueStoreItem feature
Summary:
As part of adding non-package fields to Container Manifests, we would like to store this configuration inside the image. This is so that users can have the same semantics for setting these config values as normal antlir features, with support for inheritance.

Note that we have a way to enforce ordering of these configs as well, so that users can specify ordered configs like a list.

Reviewed By: zeroxoneb

Differential Revision: D38024617

fbshipit-source-id: d00ad1c8b7511f9225a03cc75463d4922d3d5202