drop support for serde
Summary:
Supporting two formats is proving difficult, support only Thrift instead so
that we can be more careful about compatibility in saved state as well as
simplifying serialization code.

There was only one callsite using serde format anyway, so I just migrated it to
thrift, which I need to do in the next diff in this stack anyway.

Reviewed By: pallotron

Differential Revision: D38662666

fbshipit-source-id: e8c6267e8b35c5fde51242a5449d24a5bfa538f1