give Format same treatment as Kind
Summary:
When I was first writing this, it seemed like it might be useful to allow
packages of the same `Kind` to have different `Format`s, but it's since become
clearer that that is not desirable.

For example, I thought that it might be useful to allow
`SERVICE_CONFIG_GENERATOR`s to be either images or standalone files, but that
complicates other code that deals with it, so this diff moves `Format` into a
type parameter so that everything is safer.

Reviewed By: pallotron

Differential Revision: D38706240

fbshipit-source-id: 3d6cb44fe058f629d8a867fa69d680867fa3b548