rename thrift methods in preparation of async refactor
Summary:
Currently these methods are very slow and happen within the lifetime of the
single thrift request. Instead we are going to move to an async model
where the api returns a token that can be used to check on the status of the
requested operation later, instead of having to set a thrift timeout
arbitrarily high.

Reviewed By: pallotron

Differential Revision: D38353562

fbshipit-source-id: 0a7a8c88498beb1973d022ebd5be72487ec9871b