prefer binary thrift for state serialization
Summary:
Prefer binary thrift where available, fallback to json if it does not exist.

This makes it easier to maintain forward and backward compatibility with saved
state objects.

In a follow-up diff when this has rolled out everywhere, the json fallback will
be removed and instead left on disk for human consumption only.

Reviewed By: pallotron

Differential Revision: D38722993

fbshipit-source-id: 517c85ccbc839c9415d3f4faf44c4ce08e30a489