- metald library to allow import to netos metald extension
Summary:
The idea of this change is to make MetalD and the thrift server exportable and usable in a personalised version of the service.

- start_service() is the logic that can be exported and usable in any main. In our case this allow to flag metald service as netos specific or classic.
- Addressing comments from D38572714 as well: The logic for systemd is basically a symlink pointing to the netos metald version.
The interesting thing is that even if the symlink is
*/usr/lib/systemd/system/metalos.metald_test.service => /usr/lib/systemd/system/metalos_wds_metald_netos.service*
automatically the */run/systemd/system/metalos_wds_metald_netos.service* takes precedence even in the symlink, when installing the native service.

Reviewed By: pallotron

Differential Revision: D38503149

fbshipit-source-id: b3508f0363a6ca3f9458006508fd7c3cd9c3d30b