Determine disk default sector size based on fbpn list
Summary:
Determine the disk default sector size based on fbpn list.

Right now, this information is stored in the DiskSelector structure, but not used or even represented in the final output. Once T127537822 is completed, we will use this data to add a suffix to the rootdisk image.

Needs D38263166 to be landed or canaried.

Reviewed By: pzmarzly

Differential Revision: D38263368

fbshipit-source-id: d45321761517a9f742ef540d0574e6d9e4fd6f90