Consolidate storage tests and test data
Summary:
While working on adding an override library for SL, I noticed we've grown several similar-but-different hardcoded CMs used for tests, and I figured it would be worthwhile to consolidate them as our structs continue to grow, and hopefully simplify further changes + our backwards compatibility testing.

cc qhe, this may be a bit of a pain to merge with your existing stack, but the hope is it'll make future life adding more fields easier.

A summary of changes:
- Created a separate private fn for fetching the internal CM struct which will be used by override lib
- Merged private and public storage tests into one given they were covering similar functionality
- Consolidate on one source of truth struct for testing in `tupperware/cm/tests/constants.py`, use it everywhere

Reviewed By: naveedgol

Differential Revision: D38670314

fbshipit-source-id: bd090a300ed239808cab79b211376217c05458b4