[antlir] Refactor third-party build apis
Summary:
This refactors the 3rd-party antlir build APIs so that the built
artifacts are _kept_ in a resulting layer named `name` at the `/` of the
image. This means that the users of these buitl artifacts can use the
iamge layer directly for composition or extracting of the binaries.  It
doesn't imply or assume a use case for these artifacts.

Also, doing it this way makes the composition of dependencies easier
since we can just use layer mounts for pulling compiled deps into
another build (see `swtpm` for an example of this).

Reviewed By: vmagro

Differential Revision: D38246279

fbshipit-source-id: a4e96ee4a66763799e2ec30a3cecac627453354d