Allow requiring multiple keys in the meta key value store
Summary: This will be used for installing arguments and a unix_user for commands. Since these will be separate features, we need to have multiple requires. We need one require to ensure the command exists, and another require to allow chaining of arguments. For an example of how this is used check out the following diff.

Reviewed By: justintrudell

Differential Revision: D38884519

fbshipit-source-id: b48f478add41b71298033a0a8762ed84bc04d615