Add concurrency limit in update_package_db
Summary:
We have 100s of packages in the DB now, meaning this DB update creates 100s of simultaneous subprocesses shelling out to `fbpkg` when executing, which are prone to flake out.

To lazily handle this, wrap these in a semaphore so we only every have 32 futures executing at once. Not ideal performance-wise, but limits the impact for now.

Reviewed By: sergeyfd

Differential Revision: D38991617

fbshipit-source-id: f679490f50dd7ca700a92be34d9f0b333bd38055