swap: enable swap via swapfile
Summary:
# This Diff
Implements a swapfile in MetalOS. For the time being, we create a statically-sized file that's 10 * 4096 = 40KiB in size. Once swap configuration data is available in the `ProvisioningConfig`, we'll be able to set this accordingly per hostname prefix / machine type / etc. We choose a tiny value for now because we don't want to create large swapfiles inside VMtests.

Reviewed By: cooperlees

Differential Revision: D38087336

fbshipit-source-id: ba5c8e682d85ea97bcc05846d4e246732d9549db