Add remove dir to extracted features
Summary:
As requested by Servicelab, we shouldn't have to build a custom image when we're trying to remove a directory that the user added or too many jobs will use a custom image.

This adds remove to the list of extracted features. We allow the user to remove directories that were added by antlir. If it is a directory on the base image, we build a custom image.

Note that since extract features works on the pre compiler output, we just manually apply directory removal at the end

Reviewed By: epilatow

Differential Revision: D39377231

fbshipit-source-id: e3defb4d02e389fbcbeb9ff42f705e3507feca6e