only start reposerver if actually installing an rpm
Summary:
Starting the rpm snapshot server when we don't need to install any rpms is a
huge waste of time and a MASSIVE source of antlir flakiness.

Granted, we really should fix the actual flakiness, but this will go a long way
towards making signal better if we can bypass the source of flakiness in the
common case where it isn't even necessary.

Reviewed By: naveedgol

Differential Revision: D39278434

fbshipit-source-id: e9e18a5b19ba43a2c604d23ee28f243a3969e01e