Configurable, per target break points
Summary:
I discovered that the `debug-shell` that we think works, doesn't work because it doesn't actually interrupt and *stop* the boot process providing a shell on the console.  It provided the shell, but the boot just continued on making the shell useless.

# How it works
To provide an actually usable debug shell that interupts the boot process something like this approach is needed.  This `debug-shell@.service` templated service unit is enabled at build time in various `target`s.  The `target` name is the instance of the service, ie: for `sysinit.target`, `debug-shell@sysinit.service` would be enabled in the `sysinit.target`.    Then, when the VM (or machine) is booted, if the `initrd.break=<TARGET>` param is provided on the kernel cmdline, this `debug-shell@<TARGET>.service` is started, the boot process is blocked, and a shell is presented on the console.  Like so:

```
....
<console boot junk>
...
MetalOS Initrd Boot Debug Shell
--------------------------
Hostname: vmtest
Current Target: sysinit.target
--------------------------
Exit this shell to continue the boot process.
[vmtest@MetalOS Initrd ? sysinit.target]# systemctl list-jobs
JOB UNIT                   TYPE  STATE
^[[77;175R27  initrd-cleanup.service start waiting
4   debug@sysinit.service  start running
1   initrd.target          start waiting
2   basic.target           start waiting
3   sysinit.target         start waiting
25  sockets.target         start waiting
26  dbus.socket            start waiting

7 jobs listed.
[vmtest@MetalOS Initrd ? sysinit.target]# exit
...
<boot continues>
...

```

(Note: There's still some weird ansi coloring issues with the systemctl output, but I think that can be fixed up).

It is possible to provide more than one breakpoint by simply repeating the `initrd.break` cmdline argument (see the example target `//antlir/vm:default-initrd-debug`).

## Caveats

- You have to know at build time which targets you *might* want to break into.  My preference would be to simply install this for each target so they are all potential break points.  We could even provide a better systemd API to automatically do this whenever a target is installed into an `image.layer` (ie, something like `systemd.install_target(...)` that also sets up this break point service).

- This provides an *unauthenticated* root shell at boot! This is probably bad for production, but for debugging it's incredibly useful, not just for VMs but for real machines too.  One thought I had on how to handle this:  Since we have the `provisioning_config` available, we could set the root password in the initrd and use `systemd-su-login` to require password authentication to access the shell.   This would give us the protection we need but it would make it impossible to debug the process *before* acquiring the `provisioning_config` and apply the root password.  This is not as much of an issue *if* the `provsioning_config` is provided automagically to the `initrd` via concat in the bootloader.  This is something we want anyway since it removes the need for SLAAC address allocation just to retrieve the `provisioning_config`.

Reviewed By: vmagro

Differential Revision: D39323696

fbshipit-source-id: 30a978114fb07b1b3ead335574c231709fff1fa4