Tests to confirm that the VM disk image that we create will actually boot
Summary:
In the diff, we add a test to simply use antlir's VM socket to confirm that the VM booted up using the image we created will have the correct hostname.

This is a simple way to test that the image will actually boot up using `systemd` and serial console.

Reviewed By: hootio

Differential Revision: D39433818

fbshipit-source-id: 0c88f26b6c6187bbbc6b881d38c611e24796c1d8