Support skippable exceptions in update-db
Summary:
The update db script was down because a team tagged an existing pkg to point to an ephemeral, which caused the whole thing to fail.

This introduces the ability to have "skippable" exceptions in the update, meaning that failures are ignored and values are left unchanged in the DB. Given our recent frequency + the scale of updated pkgs, this is preferable to blocking the entire run.

We should also consider eventually creating corresponding tasks in this case for failures which I've linked a task in the code for.

Reviewed By: sergeyfd

Differential Revision: D39585513

fbshipit-source-id: 0207a29c5620fbc1d7e52cde386343215909d107