Update antlir rpm install to mark installed rpms
Summary:
Recently, centos8 snapshot updates were broken because `findutils` was
missing from the final image. The image construction process consisted
of:
1) install some rpms.
2) install `findutils` (and other rpms).
3) remove some rpms.

The `findutils` rpm was being removed in the third step by dnf
autoremove functionality. Dnf autoremove is enabled by default, so that
when you remove an rpm, all dependencies of that rpm will also be
removed if:
- nothing else depends on those rpms
- those rpms weren't explicitly installed by the user.

So in the image building scenario above, in step 2 we explicitly
installed `findutils`, so why is autoremove still removing it?

It turns out that `findutils` had already been installed in step 1 as
dependencies of the requested rpms. There for, in step 2 the install
requires for that rpm was a noop and dnf did not mark it as having been
requested. So it was happily removed in step 3.

To fix this, update the antlir dnf wrapper to explicitly "mark install"
all antlir requested rpms.

Note, this fix removes the workaround introduced by D39551484 to fix tw
base image builds.

Reviewed By: sergeyfd

Differential Revision: D39596860

fbshipit-source-id: b42696fa096c157a922f5644f3a12ed83858f20f