Add test utils and pretty thrift diffs
Summary:
Our Thrift structs are quite large at this point, and the failure output for any tests that compare two structs is virtually unreadable since it dumps the entire struct on left and right.

This adds a way to compare two structs and only print delta lines, then uses it in a few of our tests.

Reviewed By: sergeyfd

Differential Revision: D39701861

fbshipit-source-id: b410d540856a486f03ac9545da9ee497c2b07c61