turn off BTRFS CoW on vm disk
Summary:
TIL that `/tmp` on devservers is actually not tmpfs, but is just the root disk.
This diff creates the QCow2 image in a directory with `chattr +C` so that CoW
is disabled, which greatly increases VM disk performance.

If that still doesn't seem good enough, we can drop it in `/dev/shm` or
`/run/user/$uid` to have the disk just live in memory

Reviewed By: mzaneri

Differential Revision: D39886764

fbshipit-source-id: b11203ca2209f0d6471e4abe5656f0b96e50918d