Add a dummy rule with the label `container_manifest`
Summary:
Currently, the build service requires a target with the label `container_manifest` in order to discover all container manifest targets.

However, with a multi-variant CM, we don't have this base target, because we don't know which variant of the CM the user wants. So we create a dummy target for discoverability.

Reviewed By: justintrudell

Differential Revision: D39939763

fbshipit-source-id: 37dadd6fdfa653efc25c525712e53933d8f85a8d