Get rid of unused `layer_size_bytes` option
Summary:
This argument is never modified or used so we are basically just always passing `LOOP_SIZE` to the underlying `volume_for_repo` binary.  This removes the external API for this entirely and just hard codes the current behavior.

The use of `buck-image-out` volume is intended to be refactored pretty heavily with `buck2` native rules as well as when we move away from using a loop volume and use the native host btrfs filesystem.  So, lets get rid of this vestigle argument.

Reviewed By: vmagro

Differential Revision: D39977845

fbshipit-source-id: 415776784909641e0536ca7bafc2bdf705ae7062