remove all buck2 rules
Summary:
This is built at the wrong layer and has complicated all of antlir.
It's also responsible for all kinds of cache bugs which users have been hitting
very frequently ("feature rule does not exist" anyone?)
Leo did some great work prototyping buck2 rules in Antlir, but now we have to
rethink it to fix some of these bugs and take real advantage of what buck2
enables.

See the rest of this stack for re-implementing buck2 support from the top-down
instead of piped into the existing shitty-because-of-buck1 macros

Reviewed By: mzaneri

Differential Revision: D39973586

fbshipit-source-id: fa33c30321cbbbafd55c83d0faa275ba688cdd6b