Rebuild .btrfs package when layer changes
Summary:
The output of that `__opts` rule doesn't change when the layer changes, so Buck thinks it doesn't need to rebuild the .btrfs package.

Note that if someone wants a "for debugging" rule (this was the rationale for this intermediate rule given on the original diff) in the future, you can always reintroduce the `__opts` rule -- but don't make the `.btrfs` depend on it, that's the bug. Just use a variable to hand the same quoted JSON to both places.

Differential Revision: D40331266

fbshipit-source-id: da170b45ab86d244877501e70358bba4af88bbe1