Automatically fall back to using v1 send streams while building v2 send streams (and vice-versa)
Summary:
While rolling out v2 send streams, we will have situations where we are fetching v1
send streams during v2 builds and vice-versa.

This is a temporary workaround to ensure that we won't fail builds while we are in
this transitory phase where the old package format doesn't match the new one.

Reviewed By: snarkmaster

Differential Revision: D40479491

fbshipit-source-id: 5d160d9e22e987a7a274466c2b3478b24cb7829f