Fix recursive buck invocation for antlir/rust
Summary:
Currently, buck is emitting soft errors for when there are nested invocations with different states (i.e. the nested invocation has either a different buckconfig, or there were some source file changes). We want to fix these because these could lead to incorrect results since we are running these commands concurrently, but they have different states.

We can use `--reuse-current-config` to ensure that the same state is used for the nested invocation.

For the long term, we want to hard error recursive invocations with different states.

Sample error:

```
2022-10-10T16:21:57.057927Z  WARN buck2_core::error: Important warning at buck2/buck2_server_ctx/src/concurrency.rs:299:17 Recursive invocation of Buck, with a different state - computation will continue but may produce incorrect results. Trace Ids: 7e5e1b75-46af-4028-8850-6318eca89257, a9c5a132-9621-4434-9b29-4fe566e86d86
```

Reviewed By: epilatow

Differential Revision: D40230257

fbshipit-source-id: 3c0f49629e930d2ef55ceccf7a36f4943595e8b7