Port find-built-subvol to rust via pyo3
Summary:
First part of porting package-image to rust is to swapping out find-built-subvol for a rust implementation. Note that this does not change the existing find-built-subvol api.

There are really two things in this diff, a pure rust lib, and a pyo3 extension to add to use the rust lib in find-built-subvol. These are grouped together since they use the existing testing setup.

For new functionality to rust in antlir, I'd like to enforce a pattern wherever possible to have the rust pymodule have as little of functionality as possible, mostly acting as a type-enforcing api shim to a library underneath, so that we can more easily chain together native rust functionality in the future.

As an example of this, the unpublished diff stacked on top shows how the rust-lib is easily usable standalone via cli.

I'll probably make a post on this later, but open to discussion on this.

Reviewed By: vmagro

Differential Revision: D39517829

fbshipit-source-id: 3addc48e18c17e9707d2f73057bed3f9a51de833