replace bash.bzl with a rust_binary
Summary:
Crate a binary that wraps any arbitrary subprocess to replace
`wrap_bash_build_in_common_boilerplate`.

This is required in order to ease the transition to buck2 rules, which will
invoke this directly with the binaries from a real `toolchain()` rule, instead
of `$(exe)` expansion in a `buck_genrule`.

Reviewed By: mzaneri

Differential Revision: D40762259

fbshipit-source-id: f693ef7fa719cb2e7a37d48314e265a042265da6