[antlir][bzl] type-safe flattener function
Summary:
We frequently deal with deeply-nested lists of semi-homogenous types (think:
list of lists of lists of features) at very different levels of nesting.

This is challenging to provide a correct static type for - there are two options:
1) write a type hint for this level of nesting
2) write a type-safe flattener to immediately squash user input into one flat
 list

Option 1 quickly leads to impossible-to-read type errors with any level of
nesting beyond a single inner list

Option 2 removes the ability to provide the hint on the parameter to the top
level function, but buck2 also validates return types, so as long as we use a
type-safe flattener function we still get immediate runtime type checking.

This diff implemenets option 2 in the form of `flatten.bzl`'s `typed_flattener`
function.

Test Plan:
```
❯ buck targets fbcode//antlir/... fbcode//metalos/... fbcode//netos/... fbcode//bot_generated/kernel/... fbcode//tupperware/image/...
```

Reviewed By: nickdavies

Differential Revision: D41739274

fbshipit-source-id: bcd10e4022c6b959c73ce2dcd0d8f89b6f555b07