[antlir][oxidizing] create macro to wrap an error
Summary:
If an error type implements `Into<pyo3::PyErr>` it can be used transparently
with the `?` operator.

This macro makes it convenient to create an error wrapper for Rust-defined
exceptions and return them in Python easily.
By default, they will return in the form of a plain `Exception`, but any
exception types can be used (either from `pyo3::exceptions` or a proper class
created with `pyo3::create_exception!`

Test Plan: Used in next diff

Reviewed By: epilatow

Differential Revision: D41383186

fbshipit-source-id: ac84046c3d622ec6c738e9ecb47ab22033428325