[antlir][apt] (temporarily) rip out apt/debian support
Summary:
There is something about how this brings in a dep on the manifold client that
is breaking my buck2 work (D41792903), and the debian images are still only
hand-built so are starting to run afoul of P4P, so I am just (temporarily)
ripping it out.

SCM is our friend and we can bring it back easily when we're actually ready to
use it or push Antlir OSS again.

Test Plan: waitforsandcastle

Reviewed By: sergeyfd

Differential Revision: D42269874

fbshipit-source-id: 9a86c06370e51fbcb1edc6cf313247240a2b9963