[antlir][staging][repo-server] generate dnf.conf
Summary:
Endpoint for the repo-server to provide a dnf.conf that includes all the
repositories that the repo-server knows about.
The repo proxy knows about all the repos, and it can generate a simple dnf.conf
using the URL from the full client as the base, so it does not need to be aware
of how it's going to be proxied into the build container.

Test Plan:
```
❯ buck run 'fbcode//antlir/staging/rpm/dnf2buck/demo:test-repo[serve]' -- --bind /tmp/antlir2/sock

❯ curl -v --unix-socket /tmp/antlir2/sock http://localhost:9266/yum/dnf.conf
[fb-runtime]
baseurl=http://localhost:9266/yum/fb-runtime

❯ socat TCP6-LISTEN:9866,fork UNIX-CONNECT:/tmp/antlir2/sock &
❯ curl http://localhost:9866/yum/dnf.conf
[fb-runtime]
baseurl=http://localhost:9866/yum/fb-runtime
```

Reviewed By: zeroxoneb

Differential Revision: D42537708

fbshipit-source-id: 3afc8cf335cae5b048b4abd8111d2c66d1ba8ed4