[thrift][rust] Add support for camino::Utf8PathBuf as a rust.type for strings
Summary:
This diff adds support for `camino::Utf8PathBuf` to be used as a raw rust type for strings. This allows the Rust codegen to ensure fields are paths over strings.

Why a third-party dependency instead of `std::path::PathBuf`? Because we need to consider the round-trip method. All `Strings` are valid `PathBuf`, but not all valid `PathBuf` are `String`s. As a result, if we used the standard library's `PathBuf`, we would need to decide on the behavior for invalid Strings. I don't think we should have this behavior at the Thrift level, so instead we rely on a library that ensures that the PathBufs are a subset of Strings.

Why as a raw type for `string`s instead of `binary`s? This models off the Rust standard library, and is less unexpected than on the binary type. `PathBuf::from(String)` is trivial, but `PathBuf::from(Vec<u8>)` is not.

The implementation itself permits more than just `camino::Utf8PathBuf`: In fact, in order to implement this required adding `rust.type` support to `string` types, and the implementation accepts any `TryFrom<String>` implementor. In practice, because we don't have blanket implementations for Protocol variants of `Deserialize` and `Serialize`, I don't think users can select any type they want, which I think is an intentional feature.

Because we also don't rely on any other bounds other than `TryFrom<String>`, error messages utilize `std::any::type_name` instead for semi-decent error messages.

Test Plan: New unit test

Reviewed By: jsgf

Differential Revision: D42911302

fbshipit-source-id: 32a615ef56e208f1ebfdae744096eb257ed93024