[antlir] Write JSON mount config to mount output
Summary: Though the procfs representation is helpful on-disk, it's much more cumbersome to serde in a typed language like Rust. As such, I propose we write a simple hidden file alongside the mount that contains the JSON representation which can then be accordingly loaded in Rust (next diff).

Test Plan: Green CI

Reviewed By: zeroxoneb

Differential Revision: D42353824

fbshipit-source-id: 3ecc5d5c916cfca8343d8167df08dfe74f0663c6