[antlir] Support passing path_in_repo to find_built_subvol Rust impl
Summary:
The semantics of the `path_in_repo` arg in the `find_built_subvol` Rust lib (not the Python adapter) differ from those of the Python lib, in that for the former `path_in_repo` is really assumed to be the repo root.

Matching semantics entirely here has proven difficult because `ensure_path_in_repo` requires a Python interpreter reference and falls back to `sys.argv[0]` (from Rust!), which a bunch of existing code relies on. As a stopgap to be able to call this with similar semantics from Rust, this instead adds a fourth parameter that is a genuine `path_in_repo`.

Test Plan: Green CI, added a small unit test to cover the arg exclusion logic, could eventually expand that to create a temp file and test this function fully as well if desired.

Reviewed By: epilatow

Differential Revision: D42268911

fbshipit-source-id: 4a9f0d1d7a7c262dee994f81f44f4a23cc67eb58