virt: processvm: Build test MetalOS image
Summary:
Process VMs are a way to run VMs in Tupperware, you can read more
about them here:
https://fb.workplace.com/groups/404052403734279/posts/1333351710804339

Whilst working on Process VMs, it would be really nice to have a way to use a
`buck` built disk image for testing. In order to do that, this diff creates a
buck target that builds a simple MetalOS image with `cloud-init` installed.

I'm not really tied to using MetalOS - I just want a CentOS image that has
`cloud-init` installed.

This diff is based on the work done for the blockchain custody image, which is
located at `fbcode/blockchain/custody/vault/os/`.

Test Plan: I've tested this change with Process VMs later in the stack.

Differential Revision: D43436527

fbshipit-source-id: 53c6bf27dff6b4c8c0f6507022127581afa50d8a