[antlir] Unbundle rpm snapshots from BA images
Summary:
Today, when updating rpm snapshots, we rebuild the BA with the latest
snapshot and bundle that udpate into the snapshot itself. This is
unnecessary and we'd like to treat the BA like a regular image that gets
updated independently of the RPM snapshot. To do this we need to
unbunble the RPM snapshot from the BA. This means when we go to do a
build and download the latest BA, we should install the latest rpm
snapshot (which may not match the rpms in the BA) before using that BA
to build images.

For now just enable this new behavior for centos9-aarch64 and "-untested"
flavors.

Test Plan:
Sanity tests:
```
buck2 test \
    -c antlir.use_rc_ba=true \
    tupperware/image/build_appliance/impl/... \
    antlir/rpm/facebook/... \
    -- --run-slow-tests
```

tests_are_green_diff_is_clean

Reviewed By: zeroxoneb

Differential Revision: D43584390

fbshipit-source-id: a08a7f025e13471e2bc188c991bf01aac8edd01f