[antlir] Move build_appliance to antlir and into per-flavor directories
Summary:
This is a continuation of D42767709. In that diff we created per-flavor
directories for BA targets. Unfortunately, all the public BA targets are
still in a shared directory. This makes it impossible to create
contbuilds with different owners to cover different BAs. With the
introduction of the ELN flavor (D43643501), we now have flavors that are
owned by other oncalls, so we'd like to have them own the contbuilds for
artifacts (including the BA) that use those flavors.

Test Plan:
Sanity checks
```
buck2 cquery 'deps(//antlir/...) + deps(//tupperware/...)'
```

tests_are_green_diff_is_clean

Reviewed By: zeroxoneb

Differential Revision: D43715769

fbshipit-source-id: 898a92da46fca17125c014b3e69be7801d17ccb2