[antlir][vmtests] Allow more storage devices
Summary:
Currently we attach disks to the default PCI bus. This has a maximum of 32 slots, 4 of which are used by qemu, 1 by the network, 1 by the root disk. This means we can't have more than 26 storage devices.

Host mgmt want to test on T5-like VMs, with 36 HDDs, and this doesn't currently work.

This diff extends the vm test framework to create a PCI bridge for each set of 32 drives, and attach storage devices to that bridge instead of the default. This should increase the maximum number of storage devices to 26*32.

Test Plan: "many disks" unit test added in previous diff has been extended to use 36 drives.

Reviewed By: zeroxoneb

Differential Revision: D44254520

fbshipit-source-id: 31390befd47d14bf0e8a82d2efb335ea451af6cc