[antlir2][buck] separate transitive set types, explicit dedupe
Summary:
I need the set to be the individual artifacts, otherwise we end up with an
explosion of args when doing things like cloning many files from a single
source layer. In that case, the source layer must be the value of the tset in
order to be properly deduplicated.

I was wrong, and transitive sets do not really operate as sets in the
traditional sense, so I have to do some manual flattening here to avoid passing
the same layer path every single time a single file is cloned from it

See discussion here
https://fb.workplace.com/groups/347532827186692/posts/634411341832171

Test Plan:
```
❯ buck test fbcode//antlir/antlir2/...
```
https://www.internalfb.com/intern/testinfra/testrun/4503599794024028

Reviewed By: epilatow

Differential Revision: D45273262

fbshipit-source-id: 78ebc276aaf2d00a56355c1bd975c17ae09a074b