[antlir2][antlir2_isolate] allow registering isolated environment
Summary:
Sometimes it's necessary to get a shell inside the container for
debugging purpose. Allow registration when creating the isolation. For
Linux, this enables user to run machinectl to get a shell inside container. In
addition, this enables manual termination of the container even if the program
is not responding to interrupts.

However, as previously discussed, we don't want to expose this in the API. This
is strictly for debugging purpose.

Test Plan: buck test antlir/antlir2/test_images/mount/...

Reviewed By: vmagro

Differential Revision: D45054529

fbshipit-source-id: 3dcec6b551eb0913054ed9f91edacdbaaf67348c