[antlir2][buck] make runnable `nspawn` subtarget
Summary:
The correct semantics of this are not super clear yet, but it's useful to have
a quick way to run something in a container as a human.

This must not be used by any automation.

NOTE: since this just dumps out an invocation to `systemd-nspawn`, it doesn't
mount the repo, fbcode runtime, etc. A human being can manually add that for
now. Where I've had time to think what the correct interface is here, it can be
built into a subcommand like `antlir2 enter` and just make the image layer
target itself `buck run`-able

Test Plan:
```
❯ buck2 run fbcode//antlir/antlir2/test_images/mount:mount[nspawn]
[root@subvol-compile-f46a182448e65ccd ~]# ls /mount/in/parent/
empty
```

Reviewed By: zeroxoneb

Differential Revision: D45484522

fbshipit-source-id: 7d67321f887d85413fe7b203f7d329a9a9a5e546