Revert D45754672: Multisect successfully blamed D45754672 for test or build failures
Summary:
This diff is reverting D45754672
D45754672: [antlir][vm] remove VmTap::guest_ipv6_ll by wujj123456 has been identified to be causing the following test or build failures:

Tests affected:
- [ti/platform/edgeos/base_image/kernel_test:edgeos.vm.tests - test_services_active (ti.platform.edgeos.base_image.kernel_test.test_vm_boot.BootVMTest)](https://www.internalfb.com/intern/test/562950022317002/)

Here's the Multisect link:
https://www.internalfb.com/multisect/2047592
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Test Plan: NA

Reviewed By: wujj123456

Differential Revision: D45804335

fbshipit-source-id: 675bdb82782917e31cee1110234ec888c7d6acdc