[antlir2][enablement] implicitly create mountpoint if coming from antlir1 api
Summary:
The antlir2 api requires the image author to create their mountpoint. This makes sense.
Antlir1 however would implicitly create it for the user, so the translation api
must do the same.

If we decide to make antlir2 also implicitly create the mountpoint, that is an
easy change, but for now I just want to alter the implicit translation api
rather than change a compiler concept.

Test Plan: top of stack

Reviewed By: r1mikey

Differential Revision: D45955099

fbshipit-source-id: e1bfb90da049dd81118339128d90ce29950e9931