[static listing] Disable static listing for antlir_image_test framework.
Summary:
This change is no-op right now as static listing is disabled by default.
The reason we want to do this change is we're planning to switch the default soon and have static listing be enabled by default.
We didn't test how static listing interacts with antlir_image_test framework and our tools are limited to plain python_unittest.

Test Plan: N/A

Reviewed By: fgasperij

Differential Revision: D46145291

fbshipit-source-id: 1a608e2e9ce0820e42f8856d506330390b0d66bb