[static listing] Disable static listing for vm_test framework.
Summary:
This change is no-op right now as static listing is disabled by default.
The reason we want to do this change is we're planning to switch the default soon and have static listing be enabled by default.
We didn't test how static listing interacts with vm_test framework and our tools are limited to plain python_unittest.

Test Plan: N/A

Reviewed By: fgasperij

Differential Revision: D46145292

fbshipit-source-id: 310d007c9e09f94a712fdef9ccf714dce0a3abdc