[antlir2][rpms] add rpm excludes on flavor
Summary:
When rpms are broken for whatever reason, we need a way to exclude them for all
rpm operations for a flavor.

Test Plan:
Next diff.
I really need a `build_rule_that_fails` to properly integration-test this, just
like `test_that_fails`, but that's a lot more complicated bordering on
impossible.

Reviewed By: r1mikey

Differential Revision: D46269528

fbshipit-source-id: 5e6ad8625b634b3a9f8e399f66f200b6f0c14b81