[antlir][repo-server] ignore mutable_rpm errors
Summary:
This is not a contract that antlir is responsible for maintaining. We don't
care if the upstream repos change a given NEVRA, we just need to guarantee
reproducible rpm installations for a given SCM rev.

Test Plan: wait for sandcastle to tell me what tests are failing, fix them

Reviewed By: epilatow

Differential Revision: D46293666

fbshipit-source-id: e152e600aa7d97ca34cceae48ad871d1a65a750b